+++ src/i18n/index.ts
@@ -2,7 +2,7 @@ import { format, formatDistance, Locale } from 'date-fns'
 import i18next from 'i18next'
 import BrowserLanguageDetector from 'i18next-browser-languagedetector'
 import { initReactI18next } from 'react-i18next'
-import { ucfirst } from '~constants/strings'
+import { ucfirst } from '~utils/strings'
 import { LanguagesSlice } from './languages.mjs'
 import { dateLocales, translations } from './locales'
